legal-document-analyzer/ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.24.3
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
langchain==0.0.350
langchain-community==0.0.10
python-dotenv==1.0.0
pydantic==2.5.0
uvicorn==0.24.0
fastapi==0.104.1
ollama==0.1.7
chromadb==0.4.18
pypdf2==3.0.1
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
regex==2023.10.3
